Rename EnvPair to KeyValue.
