+++ src/msg.fbs
@@ -182,10 +182,10 @@ table SetEnv {
 }
 
 table EnvironRes {
-  map: [EnvPair];
+  map: [KeyValue];
 }
 
-table EnvPair {
+table KeyValue {
   key: string;
   value: string;
 }
